# static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moneykeeper",
    "version": "0.1.0-foundation",
    "description": "Moneykeeper context-first contract mounted by the default runtime at unversioned paths."
  },
  "paths": {
...
    "/loans": {
      "get":{"operationId":"listLoans","summary":"List loans","description":"Returns borrowed and lent agreements with confirmed component balances and accounting status.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Loan agreements"},"500":{"$ref":"#/components/responses/InternalServerError"}}},
      "post":{"operationId":"openLoan","summary":"Open a loan agreement","description":"Creates immutable contractual terms and queues Ledger liability or receivable account provisioning.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"202":{"description":"Loan opening queued"},"500":{"$ref":"#/components/responses/InternalServerError"}}}
    },
...
    "/loans/{id}/term-revisions":{
      "get":{"operationId":"listLoanTermRevisions","summary":"List loan term revisions","description":"Returns the immutable contractual revision history in revision order.","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Term revisions"},"500":{"$ref":"#/components/responses/InternalServerError"}}},
      "post":{"operationId":"reviseLoanTerms","summary":"Revise loan terms","description":"Appends a version-fenced term revision without changing posted accounting.","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdempotencyKey"}],"responses":{"200":{"description":"Terms revised"},"500":{"$ref":"#/components/responses/InternalServerError"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "parameters": {
      "IdempotencyKey": {"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":1,"maxLength":200}},
...
    },
    "responses": {
...
      "InternalServerError": {"description": "Supporting context unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {"error": {"type": "string"}}
      }
    }
  }
}